ajustes na estrutura das justificativas
